Support scaling through custom metrics API
kube-prometheus jsonnet configuration is extended to
support scaling deployments through custom metrics API.
To achieve this the the `experimental/custom-metrics-api/`
setup is in integrated into the existing jsonnet setup. In
particular:
- a ClusterRole is added to provide all-access to
  'custom.metrics.k8s.io' API group,
- a ClusterRoleBinding is added to bind previous
  ClusterRole to prometheus-adapter,
- a ClusterRoleBinding is added to bind previous
  ClusterRole to kube-system horizontal-pod-autoscaler,
- new custom metrics API ApiService is registered as
  'v1beta1.custom.metrics.k8s.io' for exposing the custom
  metrics API to kubernetes.

Additionally, `custom-metrics.jsonnet` is provided for
preparing manifests that include the new setup and allow
providing prometheus-adapter rules configuration. As an
example, `prometheus-adapter-extra-conf.yaml` is provided
which demonstrated how a new prometheus recording rule can
be introduced and map to a custom metric through which HPA
can operate.